Stop counting shifts that started outside query range
Another helpful part of PagerDuty's API: if a user starts a shift at
2022-01-31 17:00, and you want all shifts from 2022-02-01, then the API
will helpfully "chop" that overnight shift and show it as starting at
midnight.

This is not what we want for our counts; we explicitly only want to
count shifts that start within the given time range.

However, this does mean that we might need a way to distinguish between
these "chopped" shifts and shifts that legitimately do start at the
given start time. In order to address this, the query submitted to the
PagerDuty API is expanded to start a bit earlier than the desired query
window.